% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{binaryFormat}
\alias{binaryFormat}
\title{Available functions and formats for binary variables}
\usage{
binaryFormat()
}
\value{
Tibble with the available binary format keys
}
\description{
Available functions and formats for binary variables
}
