% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{categoricalFormat}
\alias{categoricalFormat}
\title{Available functions and formats for categorical variables}
\usage{
categoricalFormat()
}
\value{
Tibble with the available categorical format keys
}
\description{
Available functions and formats for categorical variables
}
