% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{numericFormat}
\alias{numericFormat}
\title{Available functions and formats for numeric variables}
\usage{
numericFormat()
}
\value{
Tibble with the available numeric format keys
}
\description{
Available functions and formats for numeric variables
}
