% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCharacteristics.R
\name{tableCharacteristics}
\alias{tableCharacteristics}
\title{Function to compute a nice table one that summarises the characteristics of
by groups. Groups can be compared between them adding absolute standardized
mean differences statistics.}
\usage{
tableCharacteristics(
  x,
  groupVariable = NULL,
  referenceGroup = NULL,
  numericVariables = NA,
  numericFormat = "median [q25 - q75]",
  dateVariables = NA,
  dateFormat = "median [min - max]",
  categoricalVariables = NA,
  categoricalFormat = "count (\%)",
  binaryVariables = NA,
  binaryFormat = "count (\%)",
  otherVariables = list(),
  otherFormat = list(),
  groupNames = list(),
  order = NULL,
  bigMark = ",",
  decimalMark = ".",
  significativeDecimals = 2
)
}
\arguments{
\item{x}{Table that contains the characteristics to be summarised.}

\item{groupVariable}{Variable of x that contains the grouping. If it is NULL
No groupping is used.}

\item{referenceGroup}{Group of reference to compute the smd. If it is NULL no
reference group is picked.}

\item{numericVariables}{Variables included in the numeric group. If NA they
are detected automatically see variableTypes(). If NULL this group is not
applied.}

\item{numericFormat}{Format for the numericVariables. See numericFormats()
for the available options. More information in xxx vignette.}

\item{dateVariables}{Variables included in the date group. If NA they are
detected automatically see variableTypes(). If NULL this group is not used.}

\item{dateFormat}{Format for the numericVariables. See dateFormats()
for the available options. More information in xxx vignette.}

\item{categoricalVariables}{Variables included in the categorical group. If
NA they are detected automatically see variableTypes(). If NULL this group is
not used.}

\item{categoricalFormat}{Format for the categoricalVariables. See
categoricalFormats() for the available options. More information in xxx
vignette.}

\item{binaryVariables}{Variables included in the binary group. If NA they are
detected automatically see variableTypes(). If NULL this group is not used.}

\item{binaryFormat}{Format for the binaryVariables. See binaryFormats() for
the available options. More information in xxx vignette.}

\item{otherVariables}{List of variables included in the different groups. All
variables should share same type.}

\item{otherFormat}{List of format for each element in otherVariables. See the
different avaialble formats in: numericFormats(), dateFormats(),
categoricalFormats() and binaryFormats().}

\item{groupNames}{List that contain the name and start pattern for the
grouped variables.}

\item{order}{Order of the variables. If NULL variables appear as the column
order. If NA they are sorted alphabetically.}

\item{bigMark}{Thousands separator.}

\item{decimalMark}{Decimal separator.}

\item{significativeDecimals}{Number of significant decimals for numeric (non
integer data)}
}
\value{
Tibble with the characteristics of the different groups
}
\description{
Function to compute a nice table one that summarises the characteristics of
by groups. Groups can be compared between them adding absolute standardized
mean differences statistics.
}
