% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{variableTypes}
\alias{variableTypes}
\title{Classify the variables between 5 types: "numeric", "categorical", "binary",
"date", or NA.}
\usage{
variableTypes(x)
}
\arguments{
\item{x}{Tibble with different columns.}
}
\value{
Tibble with the variables type and classification
}
\description{
Classify the variables between 5 types: "numeric", "categorical", "binary",
"date", or NA.
}
